Spring hw 006 2

+++ src/main/java/ru/otus/Commands/AuthorCRUD.java
@@ -1,12 +1,13 @@
 package ru.otus.Commands;
 
 //import org.jetbrains.annotations.NotNull;
+
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellOption;
 import ru.otus.Cache.Cache;
 import ru.otus.Domain.Author;
-import ru.otus.Exceptions.DataBaseException;
+import ru.otus.Exceptions.DBException;
 import ru.otus.Exceptions.NotFoundException;
 import ru.otus.Managers.AuthorManager;
 
@@ -53,22 +54,24 @@ public class AuthorCRUD {
 	@ShellMethod("Update Author by index")
 	public void updateAuthor(int index, @ShellOption(defaultValue = "")String name) {
 		Author author = (Author)cache.get(Author.class, index);
-		try {
+		if (author != null) {
 			author.setName(name);
-			authorManager.update(author);
+			author = authorManager.update(author);
 			cache.deleteAll(Author.class);
-		} catch (DataBaseException e) {
-			System.out.println(e.getMessage());
+			cache.add(Author.class, Collections.singletonList(author));
+			printAuthor(author);
 		}
 	}
 
 	@ShellMethod("Delete Author by index")
 	public void deleteAuthor(int index) {
-		Author author = (Author)cache.get(Author.class, index);
 		try {
-			authorManager.delete(author);
-			cache.delete(Author.class, index);
-		} catch (DataBaseException e) {
+			Author author = (Author) cache.get(Author.class, index);
+			if (author != null) {
+				authorManager.delete(author);
+				cache.delete(Author.class, index);
+			}
+		} catch (DBException e) {
 			System.out.println(e.getMessage());
 		}
 	}
