+++ src/main/java/ru/otus/Commands/GenreCRUD.java
@@ -1,16 +1,20 @@
 package ru.otus.Commands;
 
 //import org.jetbrains.annotations.NotNull;
+
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellOption;
 import ru.otus.Cache.Cache;
 import ru.otus.Domain.Genre;
-import ru.otus.Exceptions.DataBaseException;
+import ru.otus.Exceptions.DBException;
 import ru.otus.Exceptions.NotFoundException;
 import ru.otus.Managers.GenreManager;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 
 @ShellComponent
 public class GenreCRUD {
@@ -49,25 +53,27 @@ public class GenreCRUD {
 	@ShellMethod("Update Genre by index")
 	public void updateGenre(int index, @ShellOption(defaultValue = "")String name) {
 		Genre genre = (Genre) cache.get(Genre.class, index);
-        try {
-        	genre.setName(name);
-            genreManager.update(genre);
+		if (genre != null) {
+			genre.setName(name);
+			genreManager.update(genre);
 			cache.deleteAll(Genre.class);
-        } catch (DataBaseException e) {
-            System.out.println(e.getMessage());
-        }
+			cache.add(Genre.class, Collections.singletonList(genre));
+			printGenre(genre);
+		}
     }
 
 
 	@ShellMethod("Delete genre by index")
 	public void deleteGenre(int index) {
-		Genre genre = (Genre)cache.get(Genre.class, index);
-        try {
-            genreManager.delete(genre);
-            cache.delete(Genre.class, index);
-        } catch (DataBaseException e) {
-            System.out.println(e.getMessage());
-        }
+		try {
+			Genre genre = (Genre) cache.get(Genre.class, index);
+			if (genre != null) {
+				genreManager.delete(genre);
+				cache.delete(Genre.class, index);
+			}
+		} catch (DBException e) {
+			System.out.println(e.getMessage());
+		}
 	}
 
 	private void printGenre(Genre genre) {
